Extract shared types and css
