+++ packages/shared-components/src/room/timeline/event-tile/call/index.ts
@@ -6,3 +6,4 @@
  */
 
 export * from "./CallStartedTile/CallStartedTileView";
+export * from "./common/types";
